Created specific exception types for all generalized exceptions.  Cleaned up CLS-compliance.  Started work on the code generator for the Result enum to be used by all instance methods to return the results of the operation, instead of throwing exceptions.

+++ AW.Core/AW/Instance/AW.Instance.Functions.cs
@@ -294,10 +294,10 @@ namespace AW
             return InstanceException.Assert(InterOp.aw_terrain_set(x, z, heights.Length, texture, heights));
         }
 
-        public int TerrainQuery(int pageX, int pageZ, ulong sequence)
+        public int TerrainQuery(int pageX, int pageZ, long sequence)
         {
             SetInstance();
-            return InstanceException.Assert(InterOp.aw_terrain_query(pageX, pageZ, sequence));
+            return InstanceException.Assert(InterOp.aw_terrain_query(pageX, pageZ, (long)sequence));
         }
 
         public int TerrainNext()
